backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mediapipe
opencv-contrib-python
numpy
pillow
requests
python-dotenv